backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pdfplumber==0.11.0
python-docx==1.1.2
spacy==3.7.5
scikit-learn==1.5.1
openai==1.35.0
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
